+++ src/framework/piral-core/src/types/components.ts
@@ -189,7 +189,9 @@ export interface Errors extends PiralCustomErrors {
 /**
  * The props for the ErrorInfo component.
  */
-export type ErrorInfoProps = UnionOf<Errors>;
+export type ErrorInfoProps = UnionOf<Errors> & {
+  errorId: string;
+};
 
 /**
  * The props of a Loading indicator component.